Allow for specific `appending` and `prepending` of files
I had the intentions of adding the ability to knowingly append and prepend a stream (honoring the order of the individual stream definitions), and that's what I did. But what I didn't expect is when I first pulled down the repo, `npm test` failed. It seems that the order when the streams are combined is not preserved, and it's not predictable how they might output. It's possible a dependency (i.e. `vinyl`) changed how it behaves to remove any sort of predictability (see #3)?

Anyway, I tweaked that test and added the ability to append and prepend files and maintain order within themselves (and added tests for those and updated the README).